web/requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==3.1.1
django-constance==2.7.0
django-redis==4.12.1
django-tailwind==0.10.0
gunicorn==20.0.4
idna==2.10
isort==5.5.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.2
psycopg2==2.8.6
pylint==2.6.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
